Reorganize for easier use as library
