+++ PatientScorer.js
