+++ RankGenerator.js
